+++ tests/it/main.rs
@@ -4,6 +4,7 @@ mod buffer;
 mod ffi;
 mod scalar;
 mod temporal_conversions;
+mod types;
 
 mod io;
 mod test_util;
